Move item boolean component class to its own file
