+++ src/app/questionnaire-item/questionnaire-item.component.ts
@@ -54,23 +54,6 @@ export class QuestionnaireItemTextComponent extends QuestionnaireItemBase {
 }
 
 @Component({
-  selector: "qitem-boolean",
-  templateUrl: "./questionnaire-item-boolean.component.html",
-  styleUrls: ["./questionnaire-item.component.css"],
-})
-export class QuestionnaireItemBooleanComponent extends QuestionnaireItemBase {
-  formControl = new FormControl();
-
-  onInit() {
-    if (this.parentGroup) {
-      this.formControl = this.parentGroup.controls[
-        this.item.linkId
-      ] as FormControl;
-    }
-  }
-}
-
-@Component({
   selector: "qitem-date",
   templateUrl: "./questionnaire-item-date.component.html",
   styleUrls: ["./questionnaire-item.component.css"],
